# app/static/swagger.yaml
swagger: '2.0'
info:
  title: "Gappd Social API"
  description: "Used to manage core social features: users, posts, comments, events, photos"
  version: "0.0.1"
host: "gappd-backend.onrender.com" # During development my API is served on localhost (will change when we go to prod)
schemes:
  - "https"
consumes:
  - "application/json" # receives json data
produces:
  - "application/json" # produces json data
securityDefinitions: # Describing how to add token auth to the endpoints
  bearerAuth:
    type: apiKey
    name: Authorization
    in: header
paths: # Information about our routes
...
  /comments:
    post:
      tags:
        - Comments
      summary: "Create Comment"
      description: "Create a comment on a post; requires token"
      security:
        - bearerAuth: []
      parameters:
        - in: "body"
          name: "body"
          description: "New Comment Information"
          required: true
          schema:
            $ref: "#/definitions/CommentInput"
      responses:
        201:
          description: "Creation Successful"
          schema:
            $ref: "#/definitions/CommentResponse"
  /comments/by-post/{post_id}:
    get:
      tags:
        - Comments
      summary: "View all comments in a post"
      parameters:
        - in: "path"
          name: "post_id"
          type: "integer"
          required: true
        - in: "query"
          name: "page"
          type: "integer"
          required: false
      responses:
        200:
          description: "Successfully Retrieved Comments"
          schema:
            $ref: "#/definitions/CommentsPageResponse"
  /comments/{comment_id}:
    put:
      tags:
        - Comments
      summary: "Update Comment"
      description: "Update a comment; author only; requires token"
      security:
        - bearerAuth: []
      parameters:
        - in: "path"
          name: "comment_id"
          type: "integer"
          required: true
        - in: "body"
          name: "body"
          description: "Comment Updates"
          required: true
          schema:
            $ref: "#/definitions/CommentUpdate"
      responses:
        200:
          description: "Update Successful"
          schema:
            $ref: "#/definitions/CommentResponse"

    delete:
      tags:
        - Comments
      summary: "Delete Comment"
      description: "Delete a comment; author or owner of the parent post; requires token"
      security:
        - bearerAuth: []
      parameters:
        - in: "path"
          name: "comment_id"
          type: "integer"
          required: true
      responses:
        200:
          description: "Successfully Deleted Comment"
          schema:
            type: object
            properties:
              message:
                type: string

  #===================== Event Posts ======================
...
definitions: # Information about the shape of input and output json
...
  CommentInput:
    type: object
    properties:
      post_id:
        type: integer
        example: 1
      user_id:
        type: integer
        example: 1
      text:
        type: string
        example: "Nice bike!"
    required:
      - post_id
      - text
  CommentUpdate:
    type: object
    properties:
      text:
        type: string
  CommentResponse:
    type: object
    properties:
      id:
        type: integer
      post_id:
        type: integer
      user_id:
        type: integer
      text:
        type: string
      created_at:
        type: string
        format: date-time
  CommentsPageResponse:
    type: object
    properties:
      items:
        type: array
        items:
          $ref: "#/definitions/CommentResponse"
      page:
        type: integer
      per_page:
        type: integer
      total:
        type: integer
      pages:
        type: integer

  #==================== Event Posts & RSVP =====================
...